refactor(extraction): extract shared pipeline steps into extraction-steps.ts
Steps 2 (FormSpec generation) and 3 (AcroForm field mapping) are extracted
from createBedrockPdfExtractor into reusable helpers so both the free-JSON
and tool-use extraction variants can share them.
