+++ src/services/form-documents/extraction.ts
@@ -2,9 +2,13 @@ import { createAmazonBedrock } from '@ai-sdk/amazon-bedrock'
 import { fromNodeProviderChain } from '@aws-sdk/credential-providers'
 import { generateText } from 'ai'
 import type { CacheStore } from '../storage'
-import { enumerateFields } from './field-mapping'
-import { extractionResponseSchema, formSpecSchema } from './schemas'
-import type { ExtractionOptions, ExtractionResult, FieldMapping } from './types'
+import {
+  generateFormSpec,
+  mapAcroFormFields,
+  parseJsonResponse,
+} from './extraction-steps'
+import { extractionResponseSchema } from './schemas'
+import type { ExtractionOptions, ExtractionResult } from './types'
 
 export interface PdfExtractor {
   extract(pdf: Buffer, options?: ExtractionOptions): Promise<ExtractionResult>
@@ -46,20 +50,6 @@ export function createCachedPdfExtractor(
       return result
     },
   }
-}
-
-/** Extract JSON from a model response, stripping markdown fences if present. */
-function parseJsonResponse<T>(
-  text: string,
-  schema: { parse: (v: unknown) => T },
-): T {
-  const trimmed = text.trim()
-  // Strip ```json ... ``` fences
-  const jsonStr = trimmed.startsWith('```')
-    ? trimmed.replace(/^```(?:json)?\s*\n?/, '').replace(/\n?```\s*$/, '')
-    : trimmed
-  const parsed = JSON.parse(jsonStr)
-  return schema.parse(parsed)
 }
 
 export interface BedrockExtractorOptions {
@@ -161,82 +151,11 @@ Guidelines:
       )
 
       // Step 2: Generate default FormSpec from extracted spec
-      const formSpecResult = await generateText({
-        model: bedrock(model),
-        maxOutputTokens: 8192,
-        messages: [
-          {
-            role: 'user',
-            content: `Given this DataCollectionSpec, generate a FormSpec as JSON. Return ONLY valid JSON (no markdown, no explanation) matching this schema:
-
-{
-  "id": "form-<specId>",
-  "specId": "${spec.id}",
-  "title": "string",
-  "pages": [
-    {
-      "id": "page-1",
-      "title": "string",
-      "description": "string (optional)",
-      "groups": ["group-id-1", "group-id-2"],
-      "deliveryMode": "static|conversational|hybrid"
-    }
-  ],
-  "createdAt": "${new Date().toISOString()}",
-  "updatedAt": "${new Date().toISOString()}"
-}
-
-Rules:
-- Each page should contain 1-3 related requirement groups
-- Set deliveryMode to "static" for simple sections, "conversational" for sections with many conditional fields, "hybrid" for moderately complex sections
-
-DataCollectionSpec:
-${JSON.stringify(spec, null, 2)}`,
-          },
-        ],
-      })
-
-      const formSpec = parseJsonResponse(formSpecResult.text, formSpecSchema)
+      const bedrockModel = bedrock(model)
+      const formSpec = await generateFormSpec(bedrockModel, spec)
 
       // Step 3: Enumerate PDF AcroForm fields and map to spec fieldNames
-      const pdfFieldNames = await enumerateFields(pdf)
-      let fieldMapping: FieldMapping = {}
-
-      if (pdfFieldNames.length > 0) {
-        const allFieldNames = spec.groups
-          .flatMap((g) => g.requirements)
-          .map((r) => ({ fieldName: r.fieldName, label: r.label }))
-
-        const mappingResult = await generateText({
-          model: bedrock(model),
-          maxOutputTokens: 4096,
-          messages: [
-            {
-              role: 'user',
-              content: `Map these DataCollectionSpec fields to the PDF form fields. Return ONLY valid JSON (no markdown, no explanation) as an object where keys are spec fieldNames and values are PDF field names.
-
-Spec fields:
-${JSON.stringify(allFieldNames, null, 2)}
-
-PDF AcroForm field names:
-${JSON.stringify(pdfFieldNames, null, 2)}
-
-Rules:
-- Only include mappings where you are confident the spec field corresponds to the PDF field
-- Key = spec fieldName (camelCase), Value = exact PDF field name string
-- If a spec field has no clear PDF counterpart, omit it`,
-            },
-          ],
-        })
-
-        const mappingText = mappingResult.text.trim()
-        const jsonStr = mappingText.startsWith('```')
-          ? mappingText
-              .replace(/^```(?:json)?\s*\n?/, '')
-              .replace(/\n?```\s*$/, '')
-          : mappingText
-        fieldMapping = JSON.parse(jsonStr) as FieldMapping
-      }
+      const fieldMapping = await mapAcroFormFields(bedrockModel, pdf, spec)
 
       return {
         spec,
